+++ MainProduct.java
@@ -1,10 +1,8 @@
 package org.example;
 
-import java.math.BigDecimal;
-
 public class MainProduct extends Product{
 
-  public MainProduct(String name, BigDecimal price) {
+  public MainProduct(String name, int price) {
     super(name, price); // 親クラスのコンストラクタを呼び出す
   }
 }
